Added IWavePlayer.PlaybackTickInterval, and a running average accumulator to the EqualizerBand
This should allow you to create a viewable EQ band response. The FFT should ultimately be used; but that was another project to figure out how the frequency response relates to the EqualizerBand. The FFT window is based on a constant 1024 parameter that doesn't have anything to do with the sampling rate. So, there's more work to do to finish that part up. Then, the FFT data should be captured along with the PlaybackTick event. The additional PlaybackTickInterval throttles the PlaybackTick event to the user's liking.